feat(http-log) adds log buffering to http-log plugin
Adds three new properties to the configuration of http-log:

* `retry_count` - default 10
* `queue_size` - default 1
* `flush_timeout` - default 2

Their behaviors are the same as the identically-named options
in the `galileo` plugin.

The default of `queue_size` is set to 1 so that the default behavior
of the plugin remains backwards-compatible.

When `queue_size` is 1, it makes one HTTP request per logged
entry, containing a single JSON object in each.

When `queue_size` is greater than 1, the HTTP connection will send
a JSON array containing one or more JSON objects in it, up to
the configured queue size.

The new fields are not required, and when absent their defaults
mean that they behave the same as the http-log plugin always did,
meaning this requires no migrations.

Includes a test case using the mock log server.